requirements.txt
beautifulsoup4==4.12.3
fake-useragent==1.5.1
google-generativeai==0.8.2
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3
selenium==4.24.0
streamlit==1.40.2
